chore(auth): Move plugin options to separate files
Clean up from #2691. Moves plugin options to their own files.
