+++ packages/auth/amplify_auth_cognito_dart/lib/src/model/password/cognito_reset_password_options.dart
@@ -1,6 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+import 'package:amplify_auth_cognito_dart/amplify_auth_cognito_dart.dart';
 import 'package:amplify_core/amplify_core.dart';
 
 part 'cognito_reset_password_options.g.dart';
@@ -50,35 +51,4 @@ class CognitoResetPasswordOptions extends ResetPasswordOptions {
 
   @override
   Map<String, Object?> toJson() => _$CognitoResetPasswordOptionsToJson(this);
-}
-
-/// {@template amplify_auth_cognito.model.cognito_reset_password_plugin_options}
-/// Cognito options for `Amplify.Auth.resetPassword`.
-/// {@endtemplate}
-@zAmplifySerializable
-class CognitoResetPasswordPluginOptions extends ResetPasswordPluginOptions {
-  /// {@macro amplify_auth_cognito.model.cognito_reset_password_plugin_options}
-  const CognitoResetPasswordPluginOptions({
-    Map<String, String>? clientMetadata,
-  }) : clientMetadata = clientMetadata ?? const {};
-
-  /// {@macro amplify_auth_cognito.model.cognito_reset_password_plugin_options}
-  factory CognitoResetPasswordPluginOptions.fromJson(
-    Map<String, Object?> json,
-  ) =>
-      _$CognitoResetPasswordPluginOptionsFromJson(json);
-
-  /// Additional custom attributes to be sent to the service such as information
-  /// about the client.
-  final Map<String, String> clientMetadata;
-
-  @override
-  List<Object?> get props => [clientMetadata];
-
-  @override
-  String get runtimeTypeName => 'CognitoResetPasswordPluginOptions';
-
-  @override
-  Map<String, Object?> toJson() =>
-      _$CognitoResetPasswordPluginOptionsToJson(this);
 }
